refactor mlir compile
